requirements.txt
agno==1.4.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
docstring_parser==0.16
firecrawl-py==2.4.0
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
google-genai==1.12.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.4.3
nest-asyncio==1.6.0
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
python-dotenv==1.1.0
python-multipart==0.0.20
python-telegram-bot==22.0
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rsa==4.9.1
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
tomli==2.2.1
typer==0.15.2
typing==3.7.4.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1
yarl==1.20.0
